quell warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@19673 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/GuidedMissleStrategy.cxx
@@ -549,7 +549,7 @@ void GuidedMissileStrategy::radarRender() const
   // Display leading lines
   if (length > 0) {
     const fvec3& vel = getPath().getVelocity();
-    const fvec3 dir = vel.normalize() * shotTailLength * length;
+    const fvec3 dir = vel.normalize() * shotTailLength * (float)length;
     glBegin(GL_LINES); {
       glVertex2fv(orig);
       if (BZDBCache::leadingShotLine) {
